+++ frontend/src/components/LargeFeedCard/LargeFeedCard.styles.ts
