+++ frontend/src/components/RegularFeedCard/RegularFeedCard.styles.ts
